small changes
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@184 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ pdns/dnsproxy.cc
@@ -103,7 +103,7 @@ bool DNSProxy::sendPacket(DNSPacket *p)
   if(!recurseFor(p))
     return false;
 
-  int id;
+  u_int16_t id;
   {
     Lock l(&d_lock);
     id=getID_locked();
